requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.6.3
certifi==2026.4.22
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==47.0.0
daphne==4.2.1
Django==5.2.11
django-celery-beat==2.9.0
django-filter==25.2
django-redis==6.0.0
django-timezone-field==7.2.1
djangorestframework==3.17.1
drf-spectacular==0.29.0
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
msgpack==1.1.2
packaging==26.2
pillow==12.2.0
prompt_toolkit==3.0.52
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyOpenSSL==26.1.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qrcode==8.2
redis==7.4.0
referencing==0.37.0
requests==2.33.1
rpds-py==0.30.0
service-identity==24.2.0
setuptools==65.5.0
six==1.17.0
sqlparse==0.5.5
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
ujson==5.12.0
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.4
